+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   -------------------------------------------------------------------
 ## [Unreleased]
 - Fixed a bug in postgres where table and column names were not consistently escaped. 
+- Fixed some documentation errors.
 
 ## [1.20.0] 2021-05-06
 - Fixed a bug where postgres tables could not be anonymized if they contained a json or jsonb column.